+++ cmd/rju-dl.h
@@ -16,12 +16,12 @@ struct world {
   int32_t nc;                  /* number of channels */
   int32_t nk;                  /* number of controls */
   int32_t nb;                  /* number of buffers */
-  float sr;                    /* sample rate */
+  double sr;                   /* sample rate */
   float **in;                  /* input data */
   float **out;                 /* output data */
-  float *ctl;                  /* (shared) control data */
+  double *ctl;                 /* (shared) control data */
   float **bd;                  /* buffer data */
-  int32_t *bl;                 /* buffer sizes (0 == not-ready) */
+  int64_t *bl;                 /* buffer sizes (0 == not-ready) */
   bool ef;                     /* exit flag */
   bool vb;                     /* verbose */
 };
@@ -31,9 +31,9 @@ struct world {
 #define w_sr(w) (w)->sr
 #define w_c_get1(w,i) (w)->ctl[(i)]
 #define w_c_set1(w,i,n) (w)->ctl[(i)]=(n)
-#define w_in1(w,c,i) (w)->in[(c)][(i)]
-#define w_out1(w,c,i,n) (w)->out[(c)][(i)]=(n)
-#define w_out2(w,c,i,n1,n2) {(w)->out[(c)][(i)]=(n1);(w)->out[(c)+1][(i)]=(n2);}
+#define w_in1(w,c,i) (double)((w)->in[(c)][(i)])
+#define w_out1(w,c,i,n) (w)->out[(c)][(i)]=(float)(n)
+#define w_out2(w,c,i,n1,n2) {(w)->out[(c)][(i)]=(float)(n1);(w)->out[(c)+1][(i)]=(float)(n2);}
 /* run-time check... */
 #define w_b_read1(w,b,i) (w)->bl[(b)] > (i) ? (w)->bd[(b)][(i)] : 0.0
 #define w_b_write1(w,b,i,n) if((w)->bl[(b)] > (i)) {(w)->bd[(b)][(i)]=(n);}
